### 2024-03-05 #6444 @editorialbot
[PRE REVIEW]: EchemFEM: A Firedrake-based Python package for electrochemical transport
**Submitting author:** @tlroy (<a href="http://orcid.org/0000-0002-4286-4507">Thomas Roy</a>)
**Repository:** https://github.com/LLNL/echemfem
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.1
