### 2024-03-07 29faa2f
homog_params input to define bulk reactions
