+++ examples/carbonate_homog_params.py
@@ -1,0 +1,183 @@
+from firedrake import *
+from echemfem import EchemSolver, IntervalBoundaryLayerMesh
+
+
+class CarbonateSolver(EchemSolver):
+    """
+    Steady-state version of example from
+    Gupta, N., Gattrell, M. and MacDougall, B., 2006. Calculation for the
+    cathode surface concentrations in the electrochemical reduction of CO2 in
+    KHCO3 solutions. Journal of applied electrochemistry, 36(2), pp.161-172.
+    but using the bicarbonate bulk reactions from
+    Schulz, K.G., Riebesell, U., Rost, B., Thoms, S. and Zeebe, R.E., 2006.
+    Determination of the rate constants for the carbon dioxide to bicarbonate
+    inter-conversion in pH-buffered seawater systems. Marine chemistry,
+    100(1-2), pp.53-65.
+    """
+
+    def __init__(self):
+
+        delta = 0.0001
+        mesh = IntervalBoundaryLayerMesh(100, delta, 100, 1e-7)
+
+        # Reaction Rates in SI units
+        k1f = 2.23  # m3/mol.s
+        k1r = 9.71e-5  # 1/s
+        k2f = 3.71e-2  # 1/s
+        k2r = 2.67e1  # m3/mol.s
+        k3f = 3.06e5  # 1/s
+        k3r = 6.0e6  # m3/mol.s
+        k4f = 5.0e7  # m3/mol.s
+        k4r = 59.44  # 1/s
+        k5f = 2.31e7  # m3/mols. (Note - typo in Schulz)
+        k5r = 1.4  # mol/m3.s
+        # Bulk Conditions
+        C_HCO3_bulk = 500.  # mol/m3
+        C_CO32_bulk = 6.5  # mol/m3
+        C_OH_bulk = k3f / k3r * C_CO32_bulk / C_HCO3_bulk  # mol/m3
+        C_H_bulk = (k5r / k5f) / C_OH_bulk  # mol/m3
+        C_CO2_bulk = (k1r / k1f) * C_HCO3_bulk / C_OH_bulk  # mol/m3
+
+        homog_params = []
+
+        homog_params.append({"stoichiometry": {"CO2": -1,
+                                               "OH": -1,
+                                               "HCO3": 1},
+                             "forward rate constant": k1f,
+                             "backward rate constant": k1r
+                             })
+
+        homog_params.append({"stoichiometry": {"CO2": -1,
+                                               "H": 1,
+                                               "HCO3": 1},
+                             "forward rate constant": k2f,
+                             "backward rate constant": k2r
+                             })
+
+        homog_params.append({"stoichiometry": {"CO3": -1,
+                                               "OH": 1,
+                                               "HCO3": 1},
+                             "forward rate constant": k3f,
+                             "backward rate constant": k3r
+                             })
+
+        homog_params.append({"stoichiometry": {"CO3": -1,
+                                               "H": -1,
+                                               "HCO3": 1},
+                             "forward rate constant": k4f,
+                             "backward rate constant": k4r
+                             })
+
+        homog_params.append({"stoichiometry": {"OH": -1,
+                                               "H": -1},
+                             "forward rate constant": k5f,
+                             "backward rate constant": k5r
+                             })
+
+        conc_params = []
+
+        conc_params.append({"name": "CO2",
[…]
