### 2024-03-07 5e7157d
homog_params input to define bulk reactions
