### 2024-03-20 #6444 @RMeli
Hi @alizma 👋

Would you be interested in reviewing this submission for JOSS (Journal of Open Source Software)?

You can find more information about the submission at the top of this Github issue #6444.

> The [Journal of Open Source Software](http://joss.theoj.org/) (JOSS) is a developer friendly journal for research software packages, with a formal peer-review process that is designed to improve the quality of the software submitted.

Thank you in advance!
