### 2024-03-22 #6531 @editorialbot
[REVIEW]: EchemFEM: A Firedrake-based Python package for electrochemical transport
**Submitting author:** @tlroy (<a href="http://orcid.org/0000-0002-4286-4507">Thomas Roy</a>)
**Repository:** https://github.com/LLNL/echemfem
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.2
