### 2024-03-22 760d296
Update paper.bib with pybamm dot
