+++ joss/paper.bib
@@ -125,6 +125,7 @@
   volume={9},
   number={1},
   year={2021},
+  doi = {10.5334/jors.309},
   publisher={Ubiquity Press}
 }
 
