### 2024-03-22 e03b5d3
Update paper.bib doi for gmsh
