+++ joss/paper.bib
@@ -13,6 +13,7 @@
   number={11},
   pages={1309--1331},
   year={2009},
+  doi={10.1002/nme.2579},
   publisher={Wiley Online Library}
 }
 
