### 2024-03-24 LLNL/echemfem#8 @sudarshanv01
Installation with Firedrake
Hi @tlroy, this issue is with regards to [this](https://github.com/openjournals/joss-reviews/issues/6531#issuecomment-2016035376) review. I am having some trouble with installing `Firedrake`, which seems to be an important dependency for `echemfem`.

Please correct me if I am mistaken, but to summarize my understanding: `Firedrake` requires root access for accessing the package manager (apt/brew) in order to install required packages. There also appear to be several configuration options to install `Firedrake` through their python installer (with torch, complex, etc). I do not currently see any recommended installation instructions for `Firedrake` within the install instructions of this package. Could there be more explicit options on tags to pass to the `Firedrake` installer? Optionally, to make it easier for users, you may provide a set of required packages that can be installed by apt/brew that is required specifically for the features of `Firedrake` that this package required.

Specifically to my installation troubles, I am attempting to install `Firedrake` on a (clean) ubuntu virtual machine - allowing root access for the installation. With python3.10, the installer fails with an error that it is not able to install vtk (specifically `No matching distribution found for vtk>=9.0.1`). It seems like the only way forward to to change `firedrake-install`, are there any recommendations for what version is needed for `echemfem`?
