### 2024-03-25 LLNL/echemfem#8 @tlroy
Hi @sudarshanv01 . Firedrake is indeed crucial for EchemFEM. Installing Firedrake is not necessarily trivial as you have experienced. I am going to try to answer all your questions:
- By default apt/brew is used, but it can be disabled with `--no-package-manager`. This is usually needed when installing on clusters. Making sure all the dependencies are satisfied requires a bit of work.
- The default configuration is good for echemfem. No need for torch, complex, etc. I think most other dependencies are indeed used when running echemfem.
- I can add in the instructions that the default configuration can be used.

As for your specific installation, I have not seen this error yet. Here are some things to try:
- Is your python 32-bit or 64-bit? I think vtk only support 64-bit.
- You can attempt to fix the installation by going in the firedrake venv:
```
source firedrake/bin/activate
pip install vtk==9.0.5
firedrake-update
 ```
 - Instead of installing firedrake, you can use this docker image: https://hub.docker.com/r/firedrakeproject/firedrake
