### 2024-03-26 LLNL/echemfem#8 @sudarshanv01
Thanks for the detailed answer! The docker image works fine and I have been able to install the package. A quick follow up to close this thread - do you need editable install with your `pip install` command? An alternative that you may consider is publishing the [package to pypi](https://packaging.python.org/en/latest/tutorials/packaging-projects/).
