### 2024-03-26 LLNL/echemfem#8 @tlroy
pypi is a good idea. I will consider doing this alongside the JOSS publication.
