### 2024-04-03 LLNL/echemfem#9 @sudarshanv01
Errors in tests and memory requirements
Hello again @tlroy, I am trying to get a gauge for the memory requirements of running all the unit tests. I understand that you recommend running just one of the tests `tests/test_advection_diffusion_migration_darcy.py` instead of the whole set of tests. But when I run just `pytest` it seems to segfault, and running each unit test separately seems to give me specific errors such as
```
FAILED tests/test_advection_diffusion_migration_darcy.py::test_convergence_low_peclet - assert 0.7524928310011721 < (0.29 * 0.6053442046721568)
FAILED tests/test_advection_diffusion_migration_darcy.py::test_convergence_high_peclet - assert 0.27139794880604756 < (0.35 * 0.45044972210083223)
```
for `tests/test_advection_diffusion_migration_darcy.py`, for example.

For reference, I am running a VM with firedrake installed (as you suggested here: https://github.com/LLNL/echemfem/issues/8) on a M1 Mac.

EDIT: based on https://github.com/openjournals/joss-reviews/issues/6531#issuecomment-2016035376
