### 2024-04-03 LLNL/echemfem#8 @sudarshanv01
Sounds great! I will close this issue and tick the review for
> Installation instructions: Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.
once you let me know!
