### 2024-04-03 LLNL/echemfem#9 @tlroy
Thank you @sudarshanv01  for bringing these things to my attention.

I have not gauged the memory requirements of the tests. That is a good idea. You will need to increase the memory allocation for your docker container since the default is only 1 GB. I have mine at 16 GB but 4 GB is probably enough for the tests. Since the tests are convergence tests, they are not very fast so I recommend in the documentation to only run `tests/test_advection_diffusion_migration.py` in order to test the installation.

The Darcy tests should not actually be in the main branch. They were added by mistake and I will remove them. The Darcy and two-phase Darcy features are not full tested and are not used in any public examples or mentioned in the documentation.
