### 2024-04-03 LLNL/echemfem#10 @TomTranter
In general it would be nice to have more information about how to run the examples and how to view the output
