### 2024-04-03 LLNL/echemfem#10 @tlroy
I will keep the conversation about improving the examples in Issue #11 . I will focus here on these specific errors.
