### 2024-04-03 LLNL/echemfem#10 @tlroy
(closed by mistake)

I am not sure about the first error. It is probably due to h5py installation. Let me look into it with a fresh firedrake install.

The `paper_test.py` example will probably need to be run in parallel on most machines to get to the finest mesh. It looks like you ran out of memory. As you mentioned, the examples should be better documented so I will add something about the memory requirements of this example.

In terms of output, do you mean what is being printed? In terms of visualization, the solution fields are being stored in results/collection.pvd (and vtu files). Some 1D examples also have custom matplotlib plotting. Again, this will be better documented.
