### 2024-04-03 3f20368
Remove untested Darcy unit tests
