+++ tests/test_advection_diffusion_migration_darcy_twophase.py
@@ -1,212 +1,0 @@
-import pytest
-from firedrake import *
-from echemfem import EchemSolver
-
-
-class AdvectionDiffusionMigrationSolver(EchemSolver):
-    def __init__(self, N, vavg):
-        mesh = UnitSquareMesh(N, N, quadrilateral=True)
-        conc_params = []
-
-        x, y = SpatialCoordinate(mesh)
-        C1ex = cos(x) + sin(y) + 3
-        C2ex = C1ex
-        X1ex = 0.2 * (sin(x) + cos(y)) + 0.5
-        X2ex = 1 - X1ex
-        U1ex = sin(x) + cos(y) + 3
-        U2ex = sin(x) + cos(y) + 3
-        pex = (0.5 * cos(x) + sin(y) + 2) * vavg
-        pgex = pex
-        # pgex = cos(x) + 0.5 * sin(y) + 3
-        Sex = self.saturation(pex, pgex)
-        krl = self.relative_permeability(Sex)
-        krg = self.relative_permeability(1 - Sex)
-        vel = -krl * grad(pex)
-        velg = -krg * grad(pgex)
-        self.C1ex = C1ex
-        self.C2ex = C2ex
-        self.X1ex = X1ex
-        self.X2ex = X2ex
-        self.U1ex = U1ex
-        self.U2ex = U2ex
-        self.pex = pex
-        self.pgex = pgex
-        D1 = 0.5
-        D2 = 1.0
-        z1 = 2.0
-        z2 = -2.0
-        K = 1.0
-        Mn = 1 / (X1ex + X2ex / 2.)
-        rhog = Mn * pgex
-        sumyD = X2ex / 2.0
-        D1M = (1. - X1ex) / sumyD / Mn
-        sumyD = X1ex / 1.0
-        D2M = (1. - X2ex) / sumyD / Mn
-        D1K = 2. * 0.1 / 3. * sqrt(8. / pi / 1.)
-        D2K = 2. * 0.1 / 3. * sqrt(8. / pi / 2.)
-        DX1 = 1 / (1 / D1M + 1 / D1K)
-        DX2 = 1 / (1 / D2M + 1 / D2K)
-
-        def f(C):
-            f1 = div((vel - self.effective_diffusion(D1) * z1 * grad(U1ex)) * C1ex) - \
-                div(self.effective_diffusion(D1) * grad(C1ex))
-            f2 = div((vel - self.effective_diffusion(D2) * z2 * grad(U1ex)) * C2ex) - \
-                div(self.effective_diffusion(D2) * grad(C2ex))
-            f3 = div(- self.effective_diffusion(K, phase="solid") * grad(U2ex))
-            return [f1, f2, f3]
-
-        def fg(X):
-            f1 = div(velg * X1ex * rhog) - div(rhog * self.effective_diffusion(DX1,
-                                                                               phase="gas") * (grad(X1ex) + X1ex * grad(Mn) / Mn))
-            f2 = div(velg * X2ex * rhog) - div(rhog * self.effective_diffusion(DX2,
-                                                                               phase="gas") * (grad(X2ex) + X2ex * grad(Mn) / Mn))
-            return [f1, f2]
-
-        def fp():
-            return -div(krl * grad(pex) + D1 * z1 * grad(U1ex)
-                        * C1ex + D2 * z2 * grad(U1ex) * C2ex)
-
-        # def fpg():
-        #    return -div(krg * rhog * grad(pgex))
-
-        conc_params.append({"name": "C1",
-                            "diffusion coefficient": 0.5,
-                            "z": 2.,
-                            "bulk": C1ex,
-                            "molar mass": 1.0,
-                            })
-
-        conc_params.append({"name": "C2",
[…]
