### 2024-04-05 4e8e30f
homog_params with irreversible reactions and specified reaction order
