### 2024-04-12 5a94ff3
added example to be used as tutorial
