+++ examples/catalyst_layer_Cu_full.py
@@ -5,7 +5,12 @@ from echemfem import EchemSolver
 import numpy as np
 import csv
 """
-Model fromn
+1D model for the CO2 electrolysis in a copper catalyst layer of a gas diffusion
+electrode (GDE). The model uses electroneutral Nernst-Planck in a porous
+medium. This version of the code does not eliminate a species by substituting
+the electroneutrality equation, but rather explicitly solves for
+electroneutrality.
+
 Corral D, Lee DU, Ehlinger VM, Nitopi S, Acosta JE, Wang L, King AJ, Feaster JT, 
 Lin YR, Weber AZ, Baker SE. Bridging knowledge gaps in liquid-and vapor-fed CO2 
 electrolysis through active electrode area. Chem Catalysis. 2022 Oct 12.
@@ -240,6 +245,9 @@ for Vs in Vlist:
     solver.U_app.assign(Vs)
     print("V = %d mV" % (Vs * 1000))
     solver.solve()
+
+    ## Plotting
+
     cCO2, cOH, cH, cCO3, cHCO3, cK, phi2, phi1 = solver.u.subfunctions
 
     x_ = Function(V).interpolate(solver.mesh.coordinates[0]).vector().dat.data
