### 2024-04-22 870a048
Added why choosing Firedrake over FEniCS
