+++ joss/paper.bib
@@ -175,3 +175,38 @@
   publisher={Elsevier},
   doi={10.1016/j.carbon.2023.01.044}
 }
+
+@article{alnaes2014unified,
+  title={Unified form language: A domain-specific language for weak formulations of partial differential equations},
+  author={Aln{\ae}s, Martin S and Logg, Anders and {\O}lgaard, Kristian B and Rognes, Marie E and Wells, Garth N},
+  journal={ACM Transactions on Mathematical Software (TOMS)},
+  volume={40},
+  number={2},
+  pages={1--37},
+  year={2014},
+  publisher={ACM New York, NY, USA},
+  doi={10.1145/2566630}
+}
+
+@book{logg2012automated,
+  title={Automated solution of differential equations by the finite element method: The {FEniCS} book},
+  author={Logg, Anders and Mardal, Kent-Andre and Wells, Garth},
+  volume={84},
+  year={2012},
+  publisher={Springer Science \& Business Media},
+  doi={10.1007/978-3-642-23099-8}
+}
+
+@article{barrata2023dolfinx,
+  title={{DOLFINx}: The next generation {FEniCS} problem solving environment},
+  author={Barrata, Igor A and Dean, Joseph P and Dokken, J{\o}rgen S and HABERA, Michal and HALE, Jack and Richardson, Chris and Rognes, Marie E and Scroggs, Matthew W and Sime, Nathan and Wells, Garth N},
+  year={2023},
+  doi={10.5281/zenodo.10447666}
+}
+@article{li2024topology,
+  title={Topology Optimization for the Full-Cell Design of Porous Electrodes in Electrochemical Energy Storage Devices},
+  author={Li, Hanyu and Bucci, Giovanna and Brady, Nicholas W and Cross, Nicholas R and Ehlinger, Victoria M and Lin, Tiras Y and Salazar de Troya, Miguel and Tortorelli, Daniel and Worsley, Marcus A and Roy, Thomas},
+  journal={arXiv preprint arXiv:2403.18184},
+  year={2024},
+  doi={10.48550/arXiv.2403.18184}
+}
