+++ docs/source/demos/index.rst
@@ -1,0 +1,7 @@
+Demos
+=====
+
+.. toctree::
+   :titlesonly:
+
+   flow_reactor.py
