+++ demos/flow_reactor/flow_reactor.py.rst
@@ -1,0 +1,310 @@
+Flow reactor for CO2 electrolysis
+=================================
+
+We consider a flowing bicarbonate elecrolyte with CO2 reduction on a flat plate
+electrode.
+
+.. image:: flow_reactor.pdf
+   :width: 400px
+   :align: center
+
+
+We consider simplified CO2 - KHCO3 homogeneous reactions:
+
+.. math::
+
+   \ce{CO2} + \ce{OH-} \xrightleftharpoons[k_1^b]{k_1^f} \ce{HCO3-}
+
+   \ce{HCO3-} + \ce{OH-} \xrightleftharpoons[k_2^b]{k_2^f} \ce{CO3^2+} + \ce{H2O}
+
+Here, we assume the dilute solution case where :math:`\ce{H2O}`
+concentration is not tracked. 
+
+We are solving for :math:`c_k`, the molar concentration of species
+:math:`k\in\{\ce{CO2}, \ce{HCO3-}, \ce{CO3^2+}, \ce{OH-},
+\ce{K+}\}`,
+as well as :math:`\Phi_2`, the ionic potential.
+For each species :math:`k`, the conservation of mass is given by
+
+.. math::
+
+   \nabla \cdot \mathbf{N}_k = R_k,
+
+where :math:`\mathbf N_k` and :math:`R_k` are the flux and homogeneous
+reactions for species :math:`k`, respectively The flux term is given by the
+Nernst-Planck equation, which includes diffusion, advection, and
+electromigration:
+
+.. math::
+
+   \mathbf N_k = -D_k \nabla c_k + c_k \mathbf u - z_k F \frac{D_k}{RT} c_k \nabla \Phi_2,
+
+where :math:`D_k` is the diffusion coefficient of species :math:`k`, and
+:math:`z_k`, its charge number, :math:`\mathbf{u}` is the velocity, :math:`F`
+is the Faraday constant, :math:`R` is the ideal gas constant, :math:`T` is the
+temperature. We assume shear flow such that the velocity field is given by
+:math:`\mathbf{u} = (\dot \gamma y, 0)`, where :math:`\dot\gamma` is the shear rate and
+:math:`y` is the distance from the electrode.
+
+We assume the electroneutrality approximation such that
+
+.. math::
+
+        \sum_k z_k c_k = 0.
+
+The are five charge-transfer reactions at the surface of the electrode
+
+.. math::
+
+   \ce{CO2} + \ce{H2O} + 2\ce{e-} \rightleftharpoons \ce{HCOO-} + \ce{OH-}
+
+   \ce{CO2} + \ce{H2O} + 2\ce{e-} \rightleftharpoons \ce{CO} + 2\ce{OH-}
+
+   \ce{CO2} + 6\ce{H2O} + 8\ce{e-} \rightleftharpoons \ce{CH4} + 8\ce{OH-}
+
+   2\ce{CO2} + 8\ce{H2O} + 12\ce{e-} \rightleftharpoons \ce{C2H4} + 12\ce{OH-}
+
+   2\ce{H2O} + 2\ce{e-} \rightleftharpoons \ce{H2} + 2\ce{OH-}
+
+Note that we ignore the transport of products other than :math:`\ce{OH-}`. For each
+reaction :math:`j`, we assume a constant current density :math:`i_j`, such that
+the normal flux at the electrode surface is given by
+
+.. math::
+
+    \mathbf{N}_k \cdot \mathbf n = -\sum_j \frac{s_{k,j} i_j}{n_j F},
+
+where :math:`s_{k,j}` is the stoichiometry coefficient for the species
+:math:`k` in reaction :math:`j`, and :math:`n_j` is the number of electrons
+transferred. The partial current density is given by :math:`i_j= \mathrm{cef}_j
[…]
