+++ docs/source/quickstart.rst
@@ -7,6 +7,9 @@ Installation
 ------------
 
 First, please install the open-source finite element library `Firedrake <https://www.firedrakeproject.org/download.html>`_.
+On a Mac with Homebrew installed or on an Ubuntu workstation with sudo access, it can be installed using the default configuration.
+To get started quickly, a `Docker image <https://hub.docker.com/r/firedrakeproject/firedrake>`_ is also available.
+Note that you may need to increase the default memory allocation for Docker in order to run some of the EchemFEM examples.
 
 EchemFEM is hosted on `GitHub <https://github.com/LLNL/echemfem>`_, and should be cloned from there.
 
@@ -28,5 +31,13 @@ Alternatively, you can run some of the examples below.
 Running Examples
 ----------------
 
+To get started, a demo for a flow reactor is available :doc:`here <demos/index>`.
 Several examples can be found in `echemfem/examples <https://github.com/LLNL/echemfem/tree/main/examples>`_.
 Examples using the generalized modified Poisson-Nernst-Planck model (GMPNP) can be found in `FireCat <https://github.com/LLNL/firecat>`_, where it is coupled with a microkinetics model.
+
+Visualization
+-------------
+
+The solution fields are stored for visualization with `Paraview <https://www.paraview.org>`_ in ``results/collection.pvd``.
+Alternatively, 1D and 2D functions can be visualized using `matplotlib <https://matplotlib.org>`_. 
+More about `Firedrake visualization <https://www.firedrakeproject.org/visualisation.html>`_.
