### 2024-05-02 6152cef
minor additions to doc
