### 2024-05-02 LLNL/echemfem#10 @tlroy
I added something about memory requirements related to default docker allocation as well as information about visualization in the [QuickStart guide](https://software.llnl.gov/echemfem/quickstart.html#installation).
