### 2024-05-02 LLNL/echemfem#8 @tlroy
I added more firedrake installation details in the [QuickStart guide](https://software.llnl.gov/echemfem/quickstart.html#installation).
