### 2024-05-08 LLNL/echemfem#10 @alizma
I came across issues with some examples as well that seem to be issues with the scripts themselves rather than installation or configuration. Using the Docker linked in the documentation, the ``simple_flow_battery`` example results in the following error:
```python
Traceback (most recent call last):
  File "/home/firedrake/echemfem/examples/simple_flow_battery.py", line 117, in <module>
    solver = FlowBatterySolver()
  File "/home/firedrake/echemfem/examples/simple_flow_battery.py", line 83, in __init__
    super().__init__(
  File "/home/firedrake/echemfem/echemfem/solver.py", line 388, in __init__
    self.setup_forms(us, v)
  File "/home/firedrake/echemfem/echemfem/solver.py", line 397, in setup_forms
    self.Form, self.bcs = self.steady_forms(us, v)
  File "/home/firedrake/echemfem/echemfem/solver.py", line 460, in steady_forms
    a, bc = self.potential_poisson_form(us,
  File "/home/firedrake/echemfem/echemfem/solver.py", line 2191, in potential_poisson_form
    U_0 = self.U_app
AttributeError: 'FlowBatterySolver' object has no attribute 'U_app'
```

Additionally, when running ``catalyst_layer_Cu_full.py``, I receive the following error:
```python
Nonlinear echem_system_ solve did not converge due to DIVERGED_DTOL iterations 1
Traceback (most recent call last):
  File "/home/firedrake/echemfem/examples/catalyst_layer_Cu_full.py", line 247, in <module>
    solver.solve()
  File "/home/firedrake/echemfem/echemfem/solver.py", line 717, in solve
    self.echem_solver.solve()
  File "petsc4py/PETSc/Log.pyx", line 188, in petsc4py.PETSc.Log.EventDecorator.decorator.wrapped_func
  File "petsc4py/PETSc/Log.pyx", line 189, in petsc4py.PETSc.Log.EventDecorator.decorator.wrapped_func
  File "/home/firedrake/firedrake/src/firedrake/firedrake/adjoint_utils/variational_solver.py", line 89, in wrapper
    out = solve(self, **kwargs)
  File "/home/firedrake/firedrake/src/firedrake/firedrake/variational_solver.py", line 321, in solve
    solving_utils.check_snes_convergence(self.snes)
  File "/home/firedrake/firedrake/src/firedrake/firedrake/solving_utils.py", line 138, in check_snes_convergence
    raise ConvergenceError(r"""Nonlinear solve failed to converge after %d nonlinear iterations.
firedrake.exceptions.ConvergenceError: Nonlinear solve failed to converge after 1 nonlinear iterations.
Reason:
   DIVERGED_DTOL
```
