### 2024-05-08 LLNL/echemfem#14 @alizma
Documentation for fluid solvers?
The provided fluid solvers seem somewhat opaque and would benefit from further discussion. What are the permissible ranges of parameters? Maybe some references to how this works would also be helpful.
