### 2024-05-08 #6531 @alizma
I think the package satisfies all the technical requirements, but I'm a bit concerned with some of the examples failing and a lack of documentation for some of the solvers. @tlroy is there any chance you could take a look?
