### 2024-05-08 #6531 @RMeli
No worries @alizma.

> but I'm a bit concerned with some of the examples failing and a lack of documentation for some of the solvers

I see you have https://github.com/LLNL/echemfem/issues/14 opened for the documentation. Would you mind opening issues for the examples failing, so that the authors/developers have actionable issues to work on? Thanks.
