### 2024-05-08 1e44722
Allow zero U_app
