### 2024-05-08 LLNL/echemfem#10 @tlroy
@alizma Thanks for bringing this to my attention. I just pushed a commit to main to fix the `U_app` bug with `simple_flow_battery.py`.

For `catalyst_layer_Cu_full.py`, it looks like a difference was introduced in a recent firedrake or petsc update. On a new install (not docker), I am getting these values for the nonlinear residuals:
```
  0 SNES Function norm 2.883957200450e-03
  1 SNES Function norm 1.572419903653e+01
  2 SNES Function norm 5.269680158967e+01
Nonlinear echem_system_ solve did not converge due to DIVERGED_DTOL iterations 2
```
whereas in a docker container from March I am getting this:

```
  0 SNES Function norm 2.883957200450e-03
  1 SNES Function norm 4.878061152760e-03
  2 SNES Function norm 5.171149830053e-01
  3 SNES Function norm 1.308949961681e+01
  4 SNES Function norm 7.758895362718e+00
  5 SNES Function norm 3.502470383063e+00
  6 SNES Function norm 2.603075365207e-03
  7 SNES Function norm 1.474182974622e-03
  8 SNES Function norm 1.890120789850e-03
Nonlinear echem_system_ solve converged due to CONVERGED_SNORM_RELATIVE iterations
```
I'm asking the firedrake devs about this. In the meantime, you can get convergence by adding this line
``` solver.solver_parameters["snes_divergence_tolerance"] = -1 ```
on line 238 of the example, before the solver setup.
