### 2024-05-08 #6531 @alizma
Yes, I added the errors I ran into with examples to an existing issue about errors in the examples: https://github.com/LLNL/echemfem/issues/10, which has some discussion about fixes.
