### 2024-05-09 #6531 @RMeli
Thanks for pointing it out, I missed these additional comments on https://github.com/LLNL/echemfem/issues/10.
