### 2024-05-14 058c360
Residual weights for other equations. up divergence tolerance
