### 2024-05-14 LLNL/echemfem#10 @tlroy
@alizma So I ended up changing the values for the bulk concentrations used in `catalyst_layer_Cu_full.py`. The ones from the paper are approximations that don't solve the reaction system exactly. This means the initial guess does not solve the bulk reactions exactly and there is a mild discontinuity at the boundary, both of which could affect the convergence. So now I simply solve the system to get actual equilibrium values for the bulk concentrations. It improves the convergence.
