### 2024-05-15 ee81564
Add equations for fluid solvers in docs
