### 2024-05-15 2830dd6
more flow solver details and numerical considerations
