### 2024-05-16 LLNL/echemfem#14 @tlroy
Hi @alizma . I added some details to [the fluid solvers documentation](https://software.llnl.gov/echemfem/user_guide/fluid_solver.html), including a section at the end about the numerical method with references, and something about the range of parameters.
