### 2024-05-18 #6531 @TomTranter
Are you aware of this package? https://cidemod.readthedocs.io/en/stable/citing.html Maybe a reference to it? Could also impact statement of need
