### 2024-05-19 #6531 @tlroy
> > Are you aware of this package? https://cidemod.readthedocs.io/en/stable/citing.html Maybe a reference to it? Could also impact statement of need
>
> @TomTranter . I did not know about this package. Thanks for sharing. I will add something about it in the paper. This is actually quite similar to a battery-specific code I am working on.

Added a sentence about cideMOD in the paper.
