### 2024-05-27 #6531 @RMeli
@sudarshanv01 given your comment in https://github.com/LLNL/echemfem/issues/8#issuecomment-2094862447, do you consider the issue fixed? If yes, it can be closed.
