### 2024-05-27 #6531 @RMeli
> Thank you for the improvements and fixes. Looks good to me.

@alizma can https://github.com/LLNL/echemfem/issues/10 be closed?
