### 2024-05-27 #6531 @sudarshanv01
> @sudarshanv01 given your comment in [LLNL/echemfem#8 (comment)](https://github.com/LLNL/echemfem/issues/8#issuecomment-2094862447), do you consider the issue fixed? If yes, it can be closed.

Yup, just closed the issue.
