### 2024-05-27 #6531 @alizma
> @alizma can https://github.com/LLNL/echemfem/issues/10 be closed?

I think my concerns in that issue were addressed, but I'm not sure if the earlier comments by TomTranter were as well. If they have been, then I think the issue can be closed. Everything else looks good to me!
