+++ joss/paper.bib
@@ -210,3 +210,16 @@
   year={2024},
   doi={10.48550/arXiv.2403.18184}
 }
+
+@article{aylagas2022cidemod,
+    doi = {10.1149/1945-7111/ac91fb},
+    year = 2022,
+    month = sep,
+    publisher = {The Electrochemical Society},
+    volume = {169},
+    number = {9},
+    pages = {090528},
+    author = {Raul Ciria Aylagas and Clara Ganuza and Ruben Parra and Maria Ya{\~{n}}ez and Elixabete Ayerbe},
+    title = {{cideMOD}: An Open Source Tool for Battery Cell Inhomogeneous Performance Understanding},
+    journal = {Journal of The Electrochemical Society}
+}
