### 2024-05-31 #6531 @tlroy
@RMeli Here is the new zenodo doi: https://doi.org/10.5281/zenodo.11403579

Also, I pushed a small change to the paper in the joss branch. There was a reference missing that Tom had suggested.
