### 2024-05-31 #6531 @RMeli
> Make sure that the title and author list (including ORCIDs) in the archive match those in the JOSS paper.

> Make sure that the license listed for the archive is the same as the software license.

@tlroy Can you please edit the metadata on Zenodo to achieve the above? Thanks
