### 2024-05-31 #6531 @RMeli
Thank you @tlroy for submitting to JOSS. Based on the extensive comments from the reviewers and the review process, I'm happy to recommend this paper for acceptance! The EiC will now perform the final checks. Thank you for engaging so well with the review process.

Thank you very much again @TomTranter, @sudarshanv01, and @alizma for reviewing this submission!
