### 2024-05-31 #6531 @RMeli
Looks like the following field is creating some issues: https://github.com/LLNL/echemfem/blob/a2429c6de1a1b44b843027b0ef66b89fe778899a/joss/paper.bib#L89

Downloading the RIS data for the citation, I see this is in the SN field, which can be either ISSN, ISBN or report/document/patent number. I think it might be ISSN in this case? Probably easier to just remove it since there is a DOI and it is not a required field.
