### 2024-05-31 f3b4c0b
Update paper.bib
Remove ISBN due to joss publication error
