+++ joss/paper.bib
@@ -86,7 +86,6 @@
 @article{catmap,
 	doi = {10.1007/s10562-015-1495-6},
  	author = {Medford, Andrew J. and Shi, Chuan and Hoffmann, Max J. and Lausche, Adam C. and Fitzgibbon, Sean R. and Bligaard, Thomas and N{\o}rskov, Jens K.},
-	isbn = {1572-879X},
 	journal = {Catal. Lett.},
 	number = {3},
 	pages = {794--807},
