### 2024-05-31 #6531 @tlroy
@RMeli I removed the ISBN
