### 2024-05-31 e780093
Update paper.md
remove draft from IM number
