### 2024-05-31 #6531 @tlroy
Since this will be the published version, I just needed to remove the "DRAFT" bit of the LLNL number in the acknowledgments. Pushed a new commit. The rest looks good to me
