### 2024-05-31 #6531 @RMeli
`DRAFT` was still appearing in the latest "final" proof, but no longer appears after the latest command. Maybe I run it too early. Let's try again...
