### 2024-05-31 #6531 @Kevin-Mattheus-Moerman
@tlroy as AEiC for JOSS I will now help to process this submission for acceptance in JOSS. I have checked this review, your repository, the archive link, and the paper. All seems in order, I only have the below point that requires your attention:

- [ ] In your affiliations, please spell out USA as United States of America
